fix syntax and circular import errors
