+++ src/ixdat/backends/memory_backend.py
@@ -1,9 +1,17 @@
 """This module implements the simplest backend, which just counts rows in memory"""
-from . import BackendBase
 
 
-class MemoryBackend(BackendBase):
-    """Simplest possible backend. No saving or loading, just counting."""
+class BackendBase:
+    """Base class listing the functions that must be implemented in a database backend.
+
+    A backend defines where and how all Saveable objects (inheriting from the Saveable
+    class) save their data. This is a key part of the seamless interoperability of
+    ixdat classes and experimental databases. Each Saveable class roughly corresponds
+    to a table, and the save() and open() functions correspond to inserting and
+    selecting from a databae table.
+
+    Backends inheriting from this base class are in modules of the ixdat.backends folder
+    """
 
     def __init__(self):
         """Initialize the backend with dict for {table_name (str): id_counter (int)}"""
@@ -18,3 +26,18 @@ class MemoryBackend(BackendBase):
             i = 1
             self.next_available_ids[table_name] = 2
         return i
+
+    def save(self, obj):
+        """Save a Saveable object and return its id. Must be implemented."""
+        raise NotImplementedError
+
+    def open(self, cls, i):
+        """Load the object with id=i of a Saveable class. Must be implemented."""
+        raise NotImplementedError
+
+    def load_obj_data(self, obj):
+        """Load and return the 'data' of a saveable object. Must be implemented."""
+        raise NotImplementedError
+
+
+MemoryBackend = BackendBase
